pyproject.toml
[project]
name = "galactory"
version = "0.10.1"
authors = [
    {name = "Brian Scholer"},
]
maintainers = [
    {name = "Brian Scholer"},
]
license = {file = "LICENSE"}
description = "Ansible Galaxy proxy using Artifactory as a backend"
readme = {file = "README.md", content-type = "text/markdown"}
keywords = [
    "ansible",
    "galaxy",
    "artifactory",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"

dependencies = [
    "dohq-artifactory>=0.9,<0.10",
    "Flask>=2.1,<3",
    "semver>=2,<4",
    "base64io>=1,<2",
    "ConfigArgParse>=1.5,<2",
    "requests>=2.26.0,<3",
    "brotli>=1,<2",
]

[project.optional-dependencies]
tests = [
    "pytest>=7.1,<8",
    "pytest-cov",
    "pytest-mock",
    "coverage",
    # "pytest-docker-compose",
    # "ansible-runner",
    # "ansible-core",
]

[project.urls]
Repository = "https://github.com/briantist/galactory"
Changelog = "https://github.com/briantist/galactory/blob/main/CHANGELOG.rst"
Issues = "https://github.com/briantist/galactory/issues"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["changelogs"]

[tool.pytest.ini_options]
minversion = "7.1"
addopts = "--cov --cov-report=xml"
xfail_strict = true
# addopts = "--docker-compose=artifactory/docker-compose.yml"
# addopts = "-ra -q"
testpaths = [
    "unit",
    # "integration",
]
